现在SD卡可以被 读到，如果不开杀毒软件，速度可以达到400-500K左右

+++ Drivers/BSP/Adafruit_Shield/stm32_adafruit_sd.h
@@ -161,7 +161,7 @@ typedef struct
 {
   SD_CSD Csd;
   SD_CID Cid;
-  uint32_t CardCapacity;  /* Card Capacity */
+  uint64_t CardCapacity;  /* Card Capacity */
   uint32_t CardBlockSize; /* Card Block Size */
 } SD_CardInfo;
